Add missing interfaces
Using the snappy-debug tool to check why metrics from megaraid
are not being collected properly, the tool showed that the snap
needs hardware-observe and network-observe. However, even after
adding those interfaces the pidfd_open syscall is not allowed.
Security Engineering said that there is no interface allowing
the pidfd_open syscall ywt. They will either need to add it
to a interface if it makes sense or create a support interface